luv-test-manager: Redirect the test results to /dev/dmesg
Netconsole will only send kernel messages that are in dmesg buffer and
userspace messages are not redirected to the dmesg buffer as they can
be overwhelmingly large

In LUV it would be good to have userspace messages (messages from
running tests) included in dmesg buffer for the sake of the debugging.

Modify luv-test-manager to redirect the test results to /dev/kmsg
which will inturn inject them back to dmesg buffer. Every test result
will be provided with a timestamp which makes the job of debugging
easy in case of a system hang or panic while booting luv.

Cc:  Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Reviewed-and-tested-by: Naresh Bhat <naresh.bhat@linaro.org>
Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>