initscripts: add needed dependencies on network components
When utilizing the netconsole to debug, it becomes necessary to enable
the network interfaces if the kernel does not do this for us (e.g., when
the network driver is built as a module). Thus, add init-ifupdown as a
dependency.

Likewise, dhcp-client is added to obtain a dynamic IP address when configuring
the network interfaces.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>